Вынесли код сообщения об успешной отправке в отдельный модуль.
